Add a tox.ini
Users who want to run the tests would get this warning:

$ python3 setup.py test
running test
WARNING: Testing via this command is deprecated and will be removed in
a future version. Users looking for a generic test entry point
independent of test runner are encouraged to use tox.

Let's add a tox.ini to modernize this.